consolidated into one shared system for all four triggers
